+++ components/quiz-list/blocks/QuizListLarge.tsx
