+++ components/quiz-list/blocks/QuizListMedium.tsx
