+++ pages/challenges/[cid]/quizzes/my.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 import React from "react";
 import HeadTitle from "../../../../components/common/HeadTitle";
 import { useChallengeDetailQuery } from "../../../../api/challenges/hooks";
-import QuizListPageTemplate from "../../../../components/quizzes";
+import QuizListPageTemplate from "../../../../components/quiz-list";
 
 function MyQuizListPage() {
   const router = useRouter();
